Changed class Node to inner class of IntLinkedList
